survey: move the `sections` services under `questionnaire`
All that relates to the questionnaire should be under that service
instead of spread in multiple sub-services, as evolution is much more
than a questionnaire.
