+++ example/demo_survey/src/survey/widgets/segments.tsx
@@ -19,15 +19,15 @@ import subwayStations from '../subwayStations.json';
 import trainStations  from '../trainStations.json';
 import busRoutes  from '../busRoutes.json';
 import { GroupConfig } from 'evolution-common/lib/services/widgets';
-import { getModePreWidgetConfig } from 'evolution-common/lib/services/sections/segments/widgetSegmentModePre';
-import { getModeWidgetConfig } from 'evolution-common/lib/services/sections/segments/widgetSegmentMode';
-import { getSameAsReverseTripWidgetConfig } from 'evolution-common/lib/services/sections/segments/widgetSameAsReverseTrip';
-import { getSegmentHasNextModeWidgetConfig } from 'evolution-common/lib/services/sections/segments/widgetSegmentHasNextMode';
-import { getSegmentsGroupConfig } from 'evolution-common/lib/services/sections/segments/groupSegments';
-import { getTripSegmentsIntro } from 'evolution-common/lib/services/sections/segments/widgetTripSegmentsIntro';
-import { getButtonSaveTripSegmentsConfig } from 'evolution-common/lib/services/sections/segments/buttonSaveTripSegments';
-import { getPersonsTripsGroupConfig } from 'evolution-common/lib/services/sections/segments/groupPersonTrips';
-import { getPersonsTripsTitleWidgetConfig } from 'evolution-common/lib/services/sections/segments/widgetPersonTripsTitle';
+import { getModePreWidgetConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetSegmentModePre';
+import { getModeWidgetConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetSegmentMode';
+import { getSameAsReverseTripWidgetConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetSameAsReverseTrip';
+import { getSegmentHasNextModeWidgetConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetSegmentHasNextMode';
+import { getSegmentsGroupConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/groupSegments';
+import { getTripSegmentsIntro } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetTripSegmentsIntro';
+import { getButtonSaveTripSegmentsConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/buttonSaveTripSegments';
+import { getPersonsTripsGroupConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/groupPersonTrips';
+import { getPersonsTripsTitleWidgetConfig } from 'evolution-common/lib/services/questionnaire/sections/segments/widgetPersonTripsTitle';
 import { getFormattedDate } from 'evolution-frontend/lib/services/display/frontendHelper';
 
 export const personTrips: GroupConfig = getPersonsTripsGroupConfig();
