+++ packages/evolution-common/src/services/questionnaire/sections/common/widgetPersonVisitedPlacesMap.ts
@@ -5,13 +5,13 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 
-import { InfoMapWidgetConfig, SurveyMapObjectProperty } from '../../widgets';
+import { InfoMapWidgetConfig, SurveyMapObjectProperty } from '../../../widgets';
 
-import * as odHelpers from '../../odSurvey/helpers';
+import * as odHelpers from '../../../odSurvey/helpers';
 import { TFunction } from 'i18next';
-import projectConfig from '../../../config/project.config';
-import { pointsToBezierCurve } from '../../geodata/SurveyGeographyUtils';
-import { UserInterviewAttributes } from '../../questionnaire/types';
+import projectConfig from '../../../../config/project.config';
+import { pointsToBezierCurve } from '../../../geodata/SurveyGeographyUtils';
+import { UserInterviewAttributes } from '../../types';
 
 export const getPersonVisitedPlacesMapConfig = (
     // FIXME: Type this when there is a few more widgets implemented
