+++ packages/evolution-common/src/services/questionnaire/sections/segments/__tests__/widgetSegmentModePre.test.ts
@@ -6,11 +6,11 @@
  */
 import _cloneDeep from 'lodash/cloneDeep';
 import each from 'jest-each';
-import { InputRadioType, QuestionWidgetConfig, RadioChoiceType } from '../../../widgets';
+import { InputRadioType, QuestionWidgetConfig, RadioChoiceType } from '../../../../widgets';
 import { getModePreWidgetConfig } from '../widgetSegmentModePre';
-import { interviewAttributesForTestCases } from '../../../../tests/surveys';
-import { getResponse, setResponse, translateString } from '../../../../utils/helpers';
-import * as surveyHelper from '../../../odSurvey/helpers';
+import { interviewAttributesForTestCases } from '../../../../../tests/surveys';
+import { getResponse, setResponse, translateString } from '../../../../../utils/helpers';
+import * as surveyHelper from '../../../../odSurvey/helpers';
 import { shouldShowSameAsReverseTripQuestion, getPreviousTripSingleSegment } from '../helpers';
 
 jest.mock('../helpers', () => ({
