+++ packages/evolution-common/src/services/questionnaire/sections/segments/groupSegments.ts
@@ -5,10 +5,10 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 
-import { GroupConfig } from '../../widgets';
-import { getResponse } from '../../../utils/helpers';
+import { GroupConfig } from '../../../widgets';
+import { getResponse } from '../../../../utils/helpers';
 import { TFunction } from 'i18next';
-import { Segment } from '../../questionnaire/types';
+import { Segment } from '../../types';
 
 export const getSegmentsGroupConfig = (
     // FIXME: Type this when there is a few more widgets implemented
