+++ packages/evolution-common/src/services/questionnaire/sections/segments/widgetSameAsReverseTrip.ts
@@ -6,12 +6,12 @@
  */
 import _upperFirst from 'lodash/upperFirst';
 import { _isBlank } from 'chaire-lib-common/lib/utils/LodashExtensions';
-import { WidgetConfig } from '../../widgets';
-import { getResponse } from '../../../utils/helpers';
+import { WidgetConfig } from '../../../widgets';
+import { getResponse } from '../../../../utils/helpers';
 import { TFunction } from 'i18next';
-import * as odHelpers from '../../odSurvey/helpers';
+import * as odHelpers from '../../../odSurvey/helpers';
 import { getPreviousTripSingleSegment, shouldShowSameAsReverseTripQuestion } from './helpers';
-import { Journey, Person } from '../../questionnaire/types';
+import { Journey, Person } from '../../types';
 
 export const getSameAsReverseTripWidgetConfig = (
     // FIXME: Type this when there is a few more widgets implemented
