+++ packages/evolution-common/src/services/questionnaire/sections/segments/widgetSegmentMode.ts
@@ -6,14 +6,14 @@
  */
 import _upperFirst from 'lodash/upperFirst';
 import { _isBlank } from 'chaire-lib-common/lib/utils/LodashExtensions';
-import { WidgetConfig } from '../../widgets';
-import { getResponse } from '../../../utils/helpers';
+import { WidgetConfig } from '../../../widgets';
+import { getResponse } from '../../../../utils/helpers';
 import { TFunction } from 'i18next';
-import * as odHelpers from '../../odSurvey/helpers';
+import * as odHelpers from '../../../odSurvey/helpers';
 import { getPreviousTripSingleSegment, shouldShowSameAsReverseTripQuestion } from './helpers';
-import { Mode, modeValues } from '../../baseObjects/attributeTypes/SegmentAttributes';
-import { modePreToModeMap } from '../../odSurvey/types';
-import { ParsingFunction, Person, Segment } from '../../questionnaire/types';
+import { Mode, modeValues } from '../../../baseObjects/attributeTypes/SegmentAttributes';
+import { modePreToModeMap } from '../../../odSurvey/types';
+import { ParsingFunction, Person, Segment } from '../../types';
 
 // FIXME Move in helpers if required
 const conditionalPersonMayHaveDisability = (interview) => {
